allow delete only if all images are downloaded.

+++ Virtual Tourist/PinDetailViewController.swift
@@ -246,7 +246,9 @@ extension PinDetailViewController: UICollectionViewDelegate {
     
     func collectionView(collectionView: UICollectionView, didSelectItemAtIndexPath indexPath: NSIndexPath) {
         let photo = fetchedResultsController.objectAtIndexPath(indexPath) as! Photo
-        sharedContext.deleteObject(photo)
+        if getPhotoDownloadStatus().completed {
+            sharedContext.deleteObject(photo)
+        }
     }
 }
